requirements.txt
fastapi==0.114.0
haystack-ai==2.5.0
sentence-transformers==3.0.1
torch==2.4.1
transformers==4.44.2
uvicorn==0.30.6
accelerate==0.27.2  # Added for better model loading